Future/アクセス権限

+++ app/Http/Controllers/DiffInvitationController.php
@@ -9,6 +9,7 @@ use App\Models\Diff;
 use App\Models\UserInvitation;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
+use Gate;
 
 class DiffInvitationController extends Controller
 {
@@ -18,17 +19,20 @@ class DiffInvitationController extends Controller
      */
     public function invitations($diffId)
     {
+        Gate::authorize('diff-admin');
         $diff = Auth::user()->diffs()->findOrFail($diffId);
         $invitations = $diff->invitations()->with(['invitedPartnerUser', 'author'])->paginate();
         return Inertia::render('Diff/Invitations', [
             'invitations' => $invitations,
-            'diff' => $diff
+            'diff' => $diff,
+            'member' => $diff->findMemberByUser(Auth::user())->firstOrFail()
         ]);
     }
 
     public function search(Request $request, $diffId)
     {
         $diff = Auth::user()->diffs()->findOrFail($diffId);
+        Gate::authorize('diff-admin', $diff);
         $columns = [
             'is_member' => function($query) use ($diff){
                 $query->selectRaw('count(*)')
@@ -57,7 +61,8 @@ class DiffInvitationController extends Controller
         return Inertia::render('Diff/Invitations', [
             'users' => $page,
             'diff' => $diff,
-            'user_name' => $request->input('user_name')
+            'user_name' => $request->input('user_name'),
+            'member' => $diff->findMemberByUser(Auth::user())->firstOrFail()
         ]);
     }
 
@@ -68,6 +73,7 @@ class DiffInvitationController extends Controller
     {
         $me = Auth::user();
         $diff = $me->diffs()->findOrFail($diffId);
+        Gate::authorize('diff-admin', $diff);
         $diff->invitations()->lockForUpdate()->where('invited_partner_id','=', $userId)->delete();
         return Redirect::back()->with('success', '招待を取り下げました');
     }
@@ -77,16 +83,18 @@ class DiffInvitationController extends Controller
      * @param $diffId 招待するDiff
      * @param $userId 招待するユーザー
      */
-    public function create($diffId, $userId)
+    public function create(Request $request, $diffId, $userId)
     {
         $me = Auth::user();
         $diff = $me->diffs()->findOrFail($diffId);
+        Gate::authorize('diff-admin', $diff);
         $user = User::findOrFail($userId);
         \DB::beginTransaction();
         try{
-            $diff->invite($me, $user);
+            $diff->invite($me, $user, $request->input('authority'));
         }catch(\Exception $e){
             \DB::rollback();
+            \Log::debug($e);
             return Redirect::back()->with('error', '無効な招待です。');
         }
     
@@ -99,6 +107,7 @@ class DiffInvitationController extends Controller
     public function new(Request $request, $diffId)
     {
         $diff = Auth::user()->diffs()->findOrFail($diffId);
+        Gate::authorize('diff-admin', $diff);
         $users = User::notDiffMembers($diff)->notDiffInvitedUsers($diff)
             ->when($request->input('user_name_search') ?? null, function($query, $userNameSearch){
                 $userNameSearch = str_replace('%', '\%', $userNameSearch);
@@ -108,7 +117,8 @@ class DiffInvitationController extends Controller
         return Inertia::render('Invitation/Edit',[
             'diff' => $diff,
             'users' => $users,
-            'user_name_search' => $request->input('user_name_search')
+            'user_name_search' => $request->input('user_name_search'),
+            'member' => $diff->findMemberByUser(Auth::user())->firstOrFail()
         ]);
     }
 }
